Address PR feedback: abort guard after getSolanaKeysForRange, move defaultEnabledL2ChainIds to consts, pass explicit timeout
