+++ packages/core-mobile/app/store/network/slice.ts
@@ -14,21 +14,15 @@ import { getNetworksFromCache } from 'hooks/networks/utils/getNetworksFromCache'
 import { selectIsDeveloperMode } from 'store/settings/advanced'
 import { selectIsSolanaSupportBlocked } from 'store/posthog'
 import { selectActiveAccount } from 'store/account'
+import { defaultEnabledL2ChainIds } from 'services/network/consts'
 import { RootState } from '../types'
 import { ChainID, Networks, NetworkState } from './types'
+
+export { defaultEnabledL2ChainIds }
 
 export const defaultNetwork = BITCOIN_NETWORK
 
 export const noActiveNetwork = 0
-
-export const defaultEnabledL2ChainIds = [
-  42161, // Arbitrum One
-  // 421614, // Arbitrum Sepolia
-  10, // Optimism
-  // 11155420, // Optimism Sepolia
-  8453 // Base Mainnet
-  // 84532 // Base Sepolia
-]
 
 export const alwaysEnabledChainIds = [
   ChainsSDKChainId.AVALANCHE_MAINNET_ID,
